final commit

+++ controllers/api/comment-routes.js
@@ -1,8 +1,10 @@
 const router = require('express').Router();
 const { Comment } = require('../../models');
+const withAuth = require('../../utils/auth');
+
 
 router.get('/', (req, res) => {
-    Comment.findAll()
+  Comment.findAll()
     .then(dbCommentData => res.json(dbCommentData))
     .catch(err => {
       console.log(err);
@@ -10,14 +12,13 @@ router.get('/', (req, res) => {
     });
 });
 
-router.post('/', (req, res) => {
-  // check the session
+router.post('/', withAuth, (req, res) => {
+  // expects => {comment_text: "This is the comment", user_id: 1, post_id: 2}
   if (req.session) {
     Comment.create({
       comment_text: req.body.comment_text,
-      post_id: req.body.post_id,
-      // use the id from the session
-      user_id: req.session.user_id
+      user_id: req.session.user_id,
+      post_id: req.body.post_id
     })
       .then(dbCommentData => res.json(dbCommentData))
       .catch(err => {
@@ -27,23 +28,25 @@ router.post('/', (req, res) => {
   }
 });
 
-router.delete('/:id', (req, res) => {
+router.delete('/:id', withAuth, (req, res) => {
+  if (req.session) {
     Comment.destroy({
-        where: {
-          id: req.params.id
+      where: {
+        id: req.params.id
+      }
+    })
+      .then(dbCommentData => {
+        if (!dbCommentData) {
+          res.status(404).json({ message: 'No comment found with this id!' });
+          return;
         }
+        res.json(dbCommentData);
       })
-        .then(dbCommentData => {
-          if (!dbCommentData) {
-            res.status(404).json({ message: 'No comment found with this id' });
-            return;
-          }
-          res.json(dbCommentData);
-        })
-        .catch(err => {
-          console.log(err);
-          res.status(500).json(err);
-        });
+      .catch(err => {
+        console.log(err);
+        res.status(500).json(err);
+      });
+  }
 });
 
 module.exports = router;
